Feat: fund list 페이지 구현 완료

+++ src/main/java/fi/re/firebackend/dto/finance/fund/FundDto.java
@@ -16,5 +16,5 @@ public class FundDto {
     private String region;
     private String bseDt;
     private int selectCount;
-    private int nav;
+    private double nav;
 }
